Rework: Create new methods that output the board
Previous methods that did this were confusing and required more code to
function.

Sign arrays (previously in their own class) are now stored in the Board
class allowing those arrays to be used to create the board dynamically

Two methods were created that create each row of the board based on what
is stored in the states instance variable.

This method is called by the second method that creates the entire board

One issue:
    Empty board rows contain one less "\s" (space) character for each
    line
    This issue is not present if the "O" sign is included in the row